requirements.txt
alembic==0.7.4
beautifulsoup4==4.3.2
Flask==0.10.1
Flask-Migrate==1.3.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
psycopg2==2.6
python-dateutil==2.4.1
requests==2.6.0
six==1.9.0
SQLAlchemy==0.9.9
Werkzeug==0.10.1
